feat(component): mistakes component renamed to mistake-list component (module3-task1)
